Udemy Java

+++ UDEMY course/JAVA/ByteShortIntLong/src/academy/learnprogramming/Main.java
@@ -39,6 +39,6 @@ public class Main {
 
         byte myNewByteValue = (byte) (myMinByteValue / 2);
 
-        short myNewShortValue = (myMinShortValue / 2);
+        short myNewShortValue = (short) (myMinShortValue / 2);
     }
 }
